Fix failing unit tests after merge
These failing unit tests pointed to 2 minor bug fixes outside of test
code

The majority of failures, however, were as a result of contrived test
charms. We no longer support deploying charms which do not specify a
series or base in metadata/manifest

However, this was the case for most of our test charms. Add a
manifest.yaml to a bunch of test charms, and ensure their metadata and
manifest were correctly being returned by our mock api